backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pyodbc==5.1.0
python-multipart==0.0.9
fpdf2==2.7.9
reportlab==4.2.2
Jinja2==3.1.4
bcrypt==5.0.0
python-jose[cryptography]==3.3.0
sqlalchemy==2.0.44
aiofiles==24.1.0
playwright==1.48.0